# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SABABUKA Bersinar API
  version: 1.0.0-draft
  summary: Kontrak API MVP untuk tata kelola data Kabupaten Kapuas
  description: |
    API tunggal untuk ruang Superadmin, BAPPERIDA, Kominfo, OPD, Pimpinan,
    dashboard eksekutif, dan Asisten AI. Semua operasi dibatasi oleh permission,
    lingkup organisasi, status workflow, dan feature flag.
  contact:
    name: Pengelola SABABUKA Bersinar
  license:
    name: Penggunaan Internal Pemerintah Kabupaten Kapuas
    identifier: LicenseRef-SABABUKA-Internal
servers:
  - url: http://127.0.0.1:3001/api/v1
    description: Local development (aktif saat ini)
  - url: https://dev.sababuka.com/api/v1
    description: Development (belum diaktifkan)
  - url: https://sababuka.com/api/v1
    description: Production (setelah disetujui)
...
security:
  - sessionCookie: []
paths:
...
  /organizations:
    get:
      tags: [Identity]
      operationId: listOrganizations
      summary: Daftar organisasi
      description: Membutuhkan `organization.view`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: active
          in: query
          schema: {type: boolean}
      responses:
        "200":
          description: Daftar organisasi sesuai scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationPage"
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Identity]
      operationId: createOrganization
      summary: Membuat organisasi
      description: Membutuhkan `organization.manage`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/OrganizationInput"}
      responses:
        "201":
          description: Organisasi dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Organization"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /users:
    get:
      tags: [Identity]
      operationId: listUsers
      summary: Daftar pengguna
      description: Membutuhkan `user.view`; hasil mengikuti scope organisasi.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: organization_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
      responses:
        "200":
          description: Daftar pengguna
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: "#/components/schemas/UserSummary"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Identity]
      operationId: createUser
      summary: Membuat undangan pengguna
      description: Membutuhkan `user.create`. Pembuatan tidak otomatis memberi role.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UserCreateInput"}
      responses:
        "201":
          description: Pengguna dibuat dengan status invited
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserInvitationCreated"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /users/{user_id}/role-assignments:
    post:
      tags: [Identity]
      operationId: assignUserRole
      summary: Menetapkan role dan scope pengguna
      description: Membutuhkan `user.assign_role`; dicatat sebagai audit berisiko tinggi.
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RoleAssignmentInput"}
      responses:
        "201": {description: Assignment dibuat}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /users/{user_id}:
    patch:
      tags: [Identity]
      operationId: updateUser
      summary: Mengubah profil, status, atau kewajiban MFA pengguna
      description: Membutuhkan `user.update`; perubahan status membutuhkan `user.activate`.
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UserUpdateInput"}
      responses:
        "200":
          description: Pengguna diperbarui
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserSummary"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /roles:
    get:
      tags: [Identity]
      operationId: listRoles
      summary: Daftar role beserta permission
      description: Membutuhkan `role.view`.
      responses:
        "200":
          description: Daftar role
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Role"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /permissions:
    get:
      tags: [Identity]
      operationId: listPermissions
      summary: Daftar permission yang tersedia
      description: Membutuhkan `role.view`.
      responses:
        "200":
          description: Daftar permission
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Permission"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /roles/{role_id}/permissions:
    put:
      tags: [Identity]
      operationId: replaceRolePermissions
      summary: Mengganti permission efektif sebuah role
      description: Membutuhkan `role.manage`; permission inti Superadmin dilindungi dari penghapusan.
      parameters:
        - $ref: "#/components/parameters/RoleId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RolePermissionsInput"}
      responses:
        "200": {description: Permission role diperbarui}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /menus:
    get:
      tags: [Identity]
      operationId: listMenus
      summary: Daftar seluruh menu yang dapat dikonfigurasi
      description: Membutuhkan `menu.manage`.
      responses:
        "200":
          description: Daftar menu
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/AdminMenuItem"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /roles/{role_id}/menus:
    get:
      tags: [Identity]
      operationId: getRoleMenus
      summary: Mendapatkan menu yang tampil untuk sebuah role
      description: Membutuhkan `menu.manage`.
      parameters:
        - $ref: "#/components/parameters/RoleId"
      responses:
        "200":
          description: Identitas menu aktif untuk role
          content:
            application/json:
              schema:
                type: object
                required: [role_id, menu_ids, menu_codes]
                properties:
                  role_id: {type: string, format: uuid}
                  menu_ids:
                    type: array
                    items: {type: string, format: uuid}
                  menu_codes:
                    type: array
                    items: {type: string}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    put:
      tags: [Identity]
      operationId: replaceRoleMenus
      summary: Mengganti menu yang tampil untuk sebuah role
      description: Membutuhkan `menu.manage`; parent menu ditambahkan otomatis dan menu inti Superadmin dilindungi.
      parameters:
        - $ref: "#/components/parameters/RoleId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RoleMenusInput"}
      responses:
        "200": {description: Menu role diperbarui}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /system/configuration:
    get:
      tags: [System]
      operationId: getSystemConfiguration
      summary: Mendapatkan pengaturan nonrahasia dan feature flag
      description: Membutuhkan `system.configure`.
      responses:
        "200":
          description: Konfigurasi sistem
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SystemConfiguration"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /system/settings/{key}:
    put:
      tags: [System]
      operationId: upsertSystemSetting
      summary: Membuat atau memperbarui pengaturan nonrahasia
      description: Membutuhkan `system.configure`; secret tidak disimpan di tabel pengaturan ini.
      parameters:
        - name: key
          in: path
          required: true
          schema: {type: string, pattern: "^[a-z][a-z0-9._-]*$", maxLength: 120}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SystemSettingInput"}
      responses:
        "200": {description: Pengaturan disimpan}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /system/feature-flags/{code}:
    put:
      tags: [System]
      operationId: updateFeatureFlag
      summary: Mengubah status dan konfigurasi feature flag
      description: Membutuhkan `system.feature_manage`.
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string, pattern: "^[a-z][a-z0-9._-]*$", maxLength: 80}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FeatureFlagInput"}
      responses:
        "200": {description: Feature flag diperbarui}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sababuka_session
      description: Sesi server-side; cookie wajib HttpOnly, Secure, dan SameSite=Lax.
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 20}
      description: Token CSRF terikat ke sesi untuk seluruh operasi mutasi.
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Search:
      name: q
      in: query
      schema: {type: string, maxLength: 200}
    UserId:
      name: user_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Uuid"}
...
    RoleId:
      name: role_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Uuid"}
...
  responses:
...
    Forbidden:
      description: Permission atau scope tidak mencukupi
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    NotFound:
      description: Objek tidak ditemukan atau tidak boleh diungkapkan
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    ValidationError:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
    Conflict:
      description: Konflik kode, versi, atau data aktif
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    PageMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1}
        total_items: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {}}
        meta: {$ref: "#/components/schemas/PageMeta"}
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {$ref: "#/components/schemas/Uuid"}
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    OrganizationInput:
      type: object
      additionalProperties: false
      required: [code, name, organization_type]
      properties:
        code: {type: string, pattern: "^[A-Z0-9][A-Z0-9._-]*$", maxLength: 64}
        name: {type: string, maxLength: 255}
        short_name: {type: [string, "null"], maxLength: 120}
        organization_type: {type: string, maxLength: 40}
        parent_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
    Organization:
      allOf:
        - $ref: "#/components/schemas/OrganizationInput"
        - type: object
          required: [id, is_active, created_at, updated_at]
          properties:
            id: {$ref: "#/components/schemas/Uuid"}
            is_active: {type: boolean}
            created_at: {$ref: "#/components/schemas/Timestamp"}
            updated_at: {$ref: "#/components/schemas/Timestamp"}
    OrganizationPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/Organization"}
    UserCreateInput:
      type: object
      additionalProperties: false
      required: [email, full_name, organization_id]
      properties:
        email: {type: string, format: email, maxLength: 255}
        username: {type: [string, "null"], maxLength: 120}
        full_name: {type: string, maxLength: 255}
        organization_id: {$ref: "#/components/schemas/Uuid"}
    UserSummary:
      type: object
      required: [id, email, full_name, status]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        email: {type: string, format: email}
        username: {type: [string, "null"]}
        full_name: {type: string}
        status: {type: string, enum: [invited, active, suspended, locked, archived]}
        mfa_required: {type: boolean}
        organization_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        organization_code: {type: [string, "null"]}
        organization_name: {type: [string, "null"]}
        last_login_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        created_at: {$ref: "#/components/schemas/Timestamp"}
    UserInvitationCreated:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          required: [invitation_token, invitation_expires_at]
          properties:
            invitation_token:
              type: string
              minLength: 32
              description: Ditampilkan sekali untuk disampaikan melalui kanal internal yang aman.
            invitation_expires_at: {$ref: "#/components/schemas/Timestamp"}
    UserUpdateInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        username: {type: [string, "null"], minLength: 3, maxLength: 120}
        full_name: {type: string, minLength: 2, maxLength: 255}
        status: {type: string, enum: [invited, active, suspended, locked, archived]}
        mfa_required: {type: boolean}
    RoleAssignmentInput:
      type: object
      additionalProperties: false
      required: [role_id, scope_type]
      properties:
        role_id: {$ref: "#/components/schemas/Uuid"}
        scope_type: {type: string, enum: [global, organization, self, published]}
        organization_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        ends_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
    Role:
      type: object
      required: [id, code, name, permissions]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        name: {type: string}
        permissions:
          type: array
          items: {type: string}
        description: {type: [string, "null"]}
        is_system: {type: boolean}
        is_active: {type: boolean}
    Permission:
      type: object
      required: [id, code, name, risk_level]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        name: {type: string}
        description: {type: [string, "null"]}
        risk_level: {type: string, enum: [normal, elevated, critical]}
    RolePermissionsInput:
      type: object
      additionalProperties: false
      required: [permission_ids]
      properties:
        permission_ids:
          type: array
          uniqueItems: true
          maxItems: 500
          items: {$ref: "#/components/schemas/Uuid"}
    AdminMenuItem:
      type: object
      required: [id, code, label, display_order, is_active]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        parent_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        parent_code: {type: [string, "null"]}
        label: {type: string}
        icon: {type: [string, "null"]}
        route_name: {type: [string, "null"]}
        required_permission: {type: [string, "null"]}
        display_order: {type: integer}
        is_active: {type: boolean}
    RoleMenusInput:
      type: object
      additionalProperties: false
      required: [menu_ids]
      properties:
        menu_ids:
          type: array
          uniqueItems: true
          maxItems: 200
          items: {$ref: "#/components/schemas/Uuid"}
    SystemSettingInput:
      type: object
      additionalProperties: false
      required: [value]
      properties:
        value: {}
        description: {type: [string, "null"], maxLength: 2000}
    FeatureFlagInput:
      type: object
      additionalProperties: false
      required: [is_enabled]
      properties:
        is_enabled: {type: boolean}
        configuration: {type: object, additionalProperties: true}
    SystemConfiguration:
      type: object
      required: [settings, feature_flags]
      properties:
        settings:
          type: array
          items:
            type: object
            required: [key, value, updated_at]
            properties:
              key: {type: string}
              value: {}
              description: {type: [string, "null"]}
              updated_at: {$ref: "#/components/schemas/Timestamp"}
        feature_flags:
          type: array
          items:
            type: object
            required: [id, code, name, is_enabled, configuration, updated_at]
            properties:
              id: {$ref: "#/components/schemas/Uuid"}
              code: {type: string}
              name: {type: string}
              description: {type: [string, "null"]}
              is_enabled: {type: boolean}
              configuration: {type: object, additionalProperties: true}
              updated_at: {$ref: "#/components/schemas/Timestamp"}
...